requirements.txt
Django==1.8.4
gunicorn==19.4.5
mysqlclient==1.3.7
mysql-python==1.2.5
Pillow==3.1.1
bootstrap-admin==0.3.6
wheel==0.24.0
django-ckeditor==5.0.3
psycopg2==2.6.1
requests==2.9.1
six==1.10.0
solid-i18n==1.2.0
wsgiref==0.1.2
